Replaces connect page with interests
Migrate the connect component to being re-usable on any page. Bring in
more images to reduce text on the connect component. Create component to
wrap linkedIn button and re-use easily. Adjusts nav menu to not display
by default on small screens and to close once an item is clicked. Adds
interests component and page stub.